refactor: break up utils
